Fix:도서상세페이지 오류해결

+++ app/src/main/java/com/example/fe/search/api/BookResponse.kt
@@ -1,7 +1,9 @@
 package com.example.fe.search.api
 
+import com.google.gson.annotations.SerializedName
+
 data class BookResponse(
-    val id: Int,
+    val id: Long,
     val cover: String, // 책 표지 이미지 url
     val title: String,
     val author: String,
